Fix cases where cl-autowrap generates bindings with unusued vars
It is possible to load projects using cl-autowrap without having
loaded libffi

In those cases an error is injected into the function call to let
the user know their mistake at runtime

Becuase of injecting the error the variables holding the arguments
to the foreign function are unused and some implementatiosn emit
a style-warning

This style warning should be avoided as the primary case 'libffi is
not loaded' is covered by the error at runtime.

If failing the build is desired we should throw the error at
macroexpand-time. However I would like to keep the current beahviour
as it allows people to use the library without libffi so long as they
dont call the offending function.

This commit simply adds ignorable in the places where the local
variables could become unused. It should not cause any issue to the
usual behaviour.